requirements.txt
cloudinary==1.5.0
django-bootstrap-form==3.2.1
django-crispy-forms==1.7.0
django-markdown-deux==1.0.5
django-sekizai==0.10.0
django-test-without-migrations==0.6
dnspython==1.15.0
funcsigs==1.0.2
future==0.15.2
humanize==0.5.1
markdown2==2.3.1
mock==2.0.0
pbr==1.10.0
Pillow==2.9.0
pypiwin32==219
python-decouple==3.1
pytz==2016.6.1
singledispatch==3.4.0.3
six==1.10.0
sqlparse==0.2.2
Twisted==16.0.0
Unidecode==0.4.19
urllib3==1.19.1
zope.interface==4.3.1
graphviz==0.8.2
